Build Ubuntu 24.04 packages
Utilize `dpkg-shlibdeps(1)` to build the list of dependencies, this was needed as the old approach was not compatible with Noble.

Under Noble, we have "diversions" to handle:

```
$ dpkg -S ld-linux-aarch64.so.1
libc6:arm64: /usr/lib/aarch64-linux-gnu/ld-linux-aarch64.so.1
libc6:arm64: /usr/lib/ld-linux-aarch64.so.1
diversion by libc6 from: /lib/ld-linux-aarch64.so.1
diversion by libc6 to: /lib/ld-linux-aarch64.so.1.usr-is-merged
diversion by libc6 from: /lib/ld-linux-aarch64.so.1
diversion by libc6 to: /lib/ld-linux-aarch64.so.1.usr-is-merged
```

Using `dpkg-shlibdeps` will also make a difference in that it will add version constraints while our older approach would not. I believe it is an improvement though.

Example:

```
$ docker build --progress=plain --target tester --build-arg image=ubuntu:jammy --build-arg erlang_version=26.0 --build-arg rabbitmq_version=3.13.0 .

# now
=> Depends: procps, libc6 (>= 2.34), libgcc-s1 (>= 3.0), libstdc++6 (>= 11), libtinfo6 (>= 6), zlib1g (>= 1:1.2.8)

# before
=> Depends: procps, libc6,libgcc-s1,libstdc++6,libtinfo6,zlib1g
```