One-off build some Erlang versions missing for Noble
This is to get parity with the packages we have for Jammy and Focal.

We are missing these because they have no corresponding GitHub release at https://github.com/erlang/otp/releases.

A bit defensive settings for parallelism as it's nice to see one builds fine before having multiple failures at once :-)